Introduced EnvelopeFeatures and new operations in SamplingFeatures, that are leveraged in WaveData::sample, to make it more convenient to perform a kernel operation across an audio sample. Removed some envelope related code that will not be used.

+++ audio/src/main/java/org/almostrealism/audio/grains/GrainProcessor.java
@@ -38,7 +38,7 @@ public class GrainProcessor implements CodeFeatures {
 	}
 
 	public WaveData apply(PackedCollection<?> input, Grain grain, PackedCollection<?> wavelength, PackedCollection<?> phase) {
-		count.setMem(input.getCount());
+		count.setMem((double) input.getCount());
 		PackedCollection<?> result = ev.into(new PackedCollection<>(shape(frames), 1))
 				.evaluate(input.traverse(0), grain, wavelength, phase);
 		return new WaveData(result, sampleRate);
